Fix C compiler warnings; add long_double and short TAC cases.
Enable -Wall -Werror -Wshadow for C files (matching C++ flags); fix all
resulting warnings (const-correctness, signed/unsigned, shadow variables,
unused parameters). Add missing I16/U16 static-init and long_double
instruction cases to tac_compare and tac_graphviz.

+++ scanner/scanner.c
@@ -92,7 +92,7 @@ again:
 // Consume a character and add to yytext
 static void consume_char(void)
 {
-    if (yyleng < sizeof(yytext) - 1) {
+    if (yyleng < (int)(sizeof(yytext) - 1)) {
         yytext[yyleng++] = next_char;
         yytext[yyleng]   = '\0';
     }
